+++ ecmean/performance_indices.py
@@ -10,7 +10,6 @@
 
 import sys
 import os
-import logging
 from time import time
 from multiprocessing import Process, Manager
 import numpy as np
@@ -33,7 +32,6 @@ from ecmean.libs.parser import parse_arguments
 from ecmean.libs.loggy import setup_logger
 
 dask.config.set(scheduler="synchronous")
-
 
 class PerformanceIndices:
     """
@@ -135,9 +133,6 @@ class PerformanceIndices:
         # get file info files
         inifiles = get_inifiles(self.face, self.diag)
 
-        # add missing unit definitions
-        units_extra_definition()
-
         # create remap dictionary with atm and oce interpolators
         self.util_dictionary = Supporter(
             comp, inifiles['atm'], inifiles['oce'],
@@ -171,7 +166,8 @@ class PerformanceIndices:
         for varlist in weight_split(self.diag.field_all, self.diag.numproc):
             core = Process(target=self.pi_worker, args=(self.util_dictionary, self.piclim,
                                                         self.face, self.diag, self.diag.field_atm3d,
-                                                        self.varstat, self.outarray, varlist))
+                                                        self.varstat, self.outarray, varlist,
+                                                        self.loglevel))
             core.start()
             processes.append(core)
 
@@ -282,7 +278,7 @@ class PerformanceIndices:
 
 
     @staticmethod
-    def pi_worker(util, piclim, face, diag, field_3d, varstat, dictarray, varlist):
+    def pi_worker(util, piclim, face, diag, field_3d, varstat, dictarray, varlist, loglevel):
         """
         Main parallel diagnostic worker for performance indices.
 
@@ -296,7 +292,11 @@ class PerformanceIndices:
             dictarray (dict): Dictionary to store the output array.
             varlist (list): List of variables to process.
         """
-        loggy = logging.getLogger(__name__)
+        loggy = setup_logger(level=loglevel)
+
+        # from python 3.14 this has to be into the worker
+        units_extra_definition()
+
         for var in varlist:
             # store NaN in dict (can't use defaultdict due to multiprocessing)
             result = init_mydict(diag.seasons, diag.regions)
@@ -323,7 +323,9 @@ class PerformanceIndices:
 
                     # open file: chunking on time only, might be improved
                     if not isinstance(infile, (xr.DataArray, xr.Dataset)):
-                        xfield = xr.open_mfdataset(infile, preprocess=xr_preproc, chunks={'time': 12})
+                        xfield = xr.open_mfdataset(
+                            infile, preprocess=xr_preproc, chunks={'time': 12},
+                            data_vars='all', join='outer', compat='no_conflicts')
                     else:
                         xfield = infile