+++ CHANGELOG.md
@@ -7,8 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 Unreleased is the current development version, which is currently lying in `main` branch.
 
+- Support for python==3.14 (#136)
 - Refactored colorbar label in performance indices plots; removed kwargs direct title propagation (#133)
-
 
 ## [v0.1.16]